front: reorganize infra object layers
