+++ front/src/applications/editor/tools/selection/components/SelectionLayers.tsx
@@ -8,7 +8,7 @@ import EditorContext from 'applications/editor/context';
 import type { SelectionState } from 'applications/editor/tools/selection/types';
 import type { ExtendedEditorContextType } from 'applications/editor/types';
 import colors from 'common/Map/Consts/colors';
-import GeoJSONs from 'common/Map/Layers/GeoJSONs';
+import GeoJSONs from 'common/Map/Layers/InfraObjectLayers/GeoJSONs';
 import { useInfraID } from 'common/osrdContext';
 import { getMap } from 'reducers/map/selectors';
 import type { Zone } from 'types';
