+++ front/src/applications/editor/tools/trackEdition/components/TrackEditionLayers.tsx
@@ -10,7 +10,7 @@ import { POINTS_LAYER_ID, TRACK_LAYER_ID } from 'applications/editor/tools/track
 import type { TrackEditionState } from 'applications/editor/tools/trackEdition/types';
 import type { ExtendedEditorContextType } from 'applications/editor/types';
 import colors from 'common/Map/Consts/colors';
-import GeoJSONs from 'common/Map/Layers/GeoJSONs';
+import GeoJSONs from 'common/Map/Layers/InfraObjectLayers/GeoJSONs';
 import { useInfraID } from 'common/osrdContext';
 import { getMap } from 'reducers/map/selectors';
