+++ front/src/common/Map/Layers/InfraObjectLayers/Detectors.tsx
