+++ front/src/common/Map/Layers/InfraObjectLayers/Switches.tsx
