+++ front/src/common/Map/Layers/InfraObjectLayers/getKPLabelLayerProps.ts
