+++ front/src/common/Map/Layers/InfraObjectLayers/index.ts
@@ -1,0 +1,6 @@
+import InfraObjectLayers from './InfraObjectLayers';
+
+export { default as OperationalPointsLayer } from './OperationalPoints';
+export { default as NeutralSectionsLayer } from './extensions/SNCF/NeutralSections';
+
+export default InfraObjectLayers;
