+++ front/src/common/Map/WarpedMap/WarpedMap.tsx
@@ -11,7 +11,10 @@ import type { Layer as LayerType } from 'applications/editor/consts';
 import colors from 'common/Map/Consts/colors';
 import { ALL_SIGNAL_LAYERS } from 'common/Map/Consts/SignalsNames';
 import { useMapBlankStyle } from 'common/Map/Layers/blankStyle';
-import { EditorSource, SourcesDefinitionsIndex } from 'common/Map/Layers/GeoJSONs';
+import {
+  EditorSource,
+  SourcesDefinitionsIndex,
+} from 'common/Map/Layers/InfraObjectLayers/GeoJSONs';
 import OrderedLayer, { type OrderedLayerProps } from 'common/Map/Layers/OrderedLayer';
 import { genOSMLayerProps } from 'common/Map/Layers/OSM';
 import type { LayerContext } from 'common/Map/Layers/types';
